requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
SQLAlchemy==2.0.22
psycopg2==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4